requirements.jupyter.txt
python-dotenv
requests
ipykernel

